third_party/blink/web_tests/external/wpt/css/css-forms/parsing/picker-select-pseudo-element-expected.txt

This is a testharness.js-based test.
[FAIL] "::picker(foo)" should be an invalid selector
  assert_throws_dom: "::picker(foo)" should throw in querySelector function "() => document.querySelector(selector)" did not throw
[FAIL] "::picker(SELECT)" should be a valid selector
  assert_equals: serialization should be canonical expected "::picker(select)" but got "::picker(SELECT)"
Harness: the test ran to completion.
